Extract CLI utils module
